requirements.txt
torch-lr-finder==0.2.2
datasets==4.3.0
torch-summary==1.4.5